refactor(prover): move chain builder to ol-stf
