+++ .github/workflows/prover.yml
@@ -75,7 +75,7 @@ jobs:
           --github-token "${{ secrets.GITHUB_TOKEN }}" \
           --pr-number "${{ github.event.pull_request.number }}" \
           --commit-hash "${{ github.sha }}" \
-          --programs evm-ee-stf,cl-stf,checkpoint
+          --programs evm-ee-stf,cl-stf,checkpoint,checkpoint-new
         working-directory: bin/prover-perf
         env:
           RUSTFLAGS: "-C target-cpu=native -C link-arg=-fuse-ld=lld"
