+++ crates/proof-impl/checkpoint-new/src/lib.rs
@@ -6,8 +6,6 @@
 //! to the protocol rules.
 #[cfg(not(target_os = "zkvm"))]
 pub mod program;
-#[cfg(any(test, feature = "test-utils"))]
-pub mod test_utils;
 
 mod statements;
